+++ packages/admin-ui/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 
 -   Wrap component styles in a `wp-admin-ui` cascade layer so consumer overrides are not blocked by library specificity. ([#82080](https://github.com/WordPress/gutenberg/pull/82080)).
+-   Add unlayered wp-admin list defenses for layered `Breadcrumbs` and `Navigation` list markup. ([#82080](https://github.com/WordPress/gutenberg/pull/82080)).
 
 ### Bug Fixes